getlantern/lantern#1888 Moved client/server code into its own file
